+++ app/src/test/java/com/rootstrap/android/test/UnitTestBase.kt
@@ -1,4 +1,4 @@
-package com.rootstrap.android
+package com.rootstrap.android.test
 
 import androidx.annotation.CallSuper
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
